added remote ipfs cluster host, bit code refactoring
